# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sayenly API Documentation",
    "version": "1.0.0",
    "description": "Documentation for Sayenly Backend API",
    "contact": {
      "name": "Sayenly Support",
      "email": "support@sayenly.com"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UUID": {
        "type": "string",
        "format": "uuid",
        "description": "A Universally Unique Identifier (UUID) v4",
        "example": "123e4567-e89b-12d3-a456-426614174000"
      },
      "Email": {
        "type": "string",
        "format": "email",
        "description": "A valid email address",
        "example": "user@example.com"
      },
...
      "DateTime": {
        "type": "string",
        "format": "date-time",
        "description": "A date-time string in ISO 8601 format",
        "example": "2023-01-23T14:30:00Z"
      },
      "URL": {
        "type": "string",
        "format": "uri",
        "description": "A valid URL",
        "example": "https://example.com/resource"
      },
      "JSON": {
        "type": "object",
        "description": "A JSON object",
        "additionalProperties": true,
        "example": {
          "key": "value",
          "nested": {
            "data": 123
          }
        }
      },
      "PhoneNumber": {
        "type": "string",
        "description": "A phone number that complies with E.164 format",
        "pattern": "^\\+?[1-9]\\d{1,14}$",
        "example": "+12125551212"
      },
...
      "Company": {
        "type": "object",
        "required": [
          "id",
          "name",
          "phoneNumbers",
          "emailAddresses",
          "addresses",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "name": {
            "type": "string",
            "description": "Name of the company"
          },
          "logo": {
            "$ref": "#/components/schemas/URL",
            "description": "URL to the company's logo"
          },
          "phoneNumbers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PhoneNumber"
            },
            "description": "List of phone numbers for the company"
          },
          "emailAddresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Email"
            },
            "description": "List of email addresses for the company"
          },
          "addresses": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of physical addresses for the company"
          },
          "services": {
            "type": "array",
            "description": "Services provided by this company",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "$ref": "#/components/schemas/UUID"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          },
          "totalEarnings": {
            "type": "number",
            "description": "Total earnings for the company"
          },
          "createdAt": {
            "$ref": "#/components/schemas/DateTime"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/DateTime"
          }
        }
      },
...
      "Project": {
        "type": "object",
        "required": [
          "id",
          "headline",
          "description",
          "images",
          "address",
          "date",
          "htmlContent",
          "inApp",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "headline": {
            "type": "string",
            "description": "Main headline for the project"
          },
          "description": {
            "type": "string",
            "description": "Description of the project"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/URL"
            },
            "description": "List of image URLs for the project"
          },
          "address": {
            "type": "string",
            "description": "Location address of the project"
          },
          "date": {
            "$ref": "#/components/schemas/DateTime"
          },
          "inApp": {
            "type": "boolean",
            "description": "Whether the project is visible in the app"
          },
          "htmlContent": {
            "$ref": "#/components/schemas/JSON",
            "description": "Rich text content for the project"
          },
          "services": {
            "type": "array",
            "description": "Services related to this project",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "$ref": "#/components/schemas/UUID"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          },
          "areas": {
            "type": "array",
            "description": "Areas related to this project",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "$ref": "#/components/schemas/UUID"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          },
          "createdAt": {
            "$ref": "#/components/schemas/DateTime"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/DateTime"
          }
        }
      },
...
      "Service": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "pastJobs",
          "inApp",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The auto-generated ID of the service"
          },
          "name": {
            "type": "string",
            "description": "Name of the service"
          },
          "description": {
            "type": "string",
            "description": "Description of the service"
          },
          "pastJobs": {
            "type": "integer",
            "description": "Number of past jobs completed for this service"
          },
          "serviceCardImage": {
            "type": "string",
            "description": "URL to the service's card image"
          },
          "inApp": {
            "type": "boolean",
            "description": "Whether the service is visible in the app"
          },
          "Projects": {
            "type": "array",
            "description": "Projects related to this service",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          },
          "company": {
            "type": "array",
            "description": "Companies providing this service",
            "items": {
              "$ref": "#/components/schemas/Company"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date the service was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date the service was last updated"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/service/get-all": {
      "get": {
        "summary": "Get all services",
        "tags": [
          "Service"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all services",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Service"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/service/get-by-id": {
      "get": {
        "summary": "Get a service by ID",
        "tags": [
          "Service"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The service ID"
          }
        ],
        "responses": {
          "200": {
            "description": "The service information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/service/create": {
      "post": {
        "summary": "Create a new service",
        "tags": [
          "Service"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "description",
                  "pastJobs"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the service"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the service"
                  },
                  "pastJobs": {
                    "type": "integer",
                    "description": "Number of past jobs completed for this service"
                  },
                  "serviceCardImage": {
                    "type": "string",
                    "description": "URL to the service's card image"
                  },
                  "inApp": {
                    "type": "boolean",
                    "default": false,
                    "description": "Whether the service is visible in the app"
                  },
                  "Projects": {
                    "type": "array",
                    "description": "Projects to associate with this service",
                    "items": {
                      "type": "object",
                      "required": [
                        "id"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "description": "Project ID"
                        }
                      }
                    }
                  },
                  "company": {
                    "type": "array",
                    "description": "Companies providing this service",
                    "items": {
                      "type": "object",
                      "required": [
                        "id"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "description": "Company ID"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/service/update": {
      "put": {
        "summary": "Update a service",
        "tags": [
          "Service"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The service ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the service"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the service"
                  },
                  "pastJobs": {
                    "type": "integer",
                    "description": "Number of past jobs completed for this service"
                  },
                  "serviceCardImage": {
                    "type": "string",
                    "description": "URL to the service's card image"
                  },
                  "inApp": {
                    "type": "boolean",
                    "description": "Whether the service is visible in the app"
                  },
                  "Projects": {
                    "type": "array",
                    "description": "Projects to associate with this service",
                    "items": {
                      "type": "object",
                      "required": [
                        "id"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "description": "Project ID"
                        }
                      }
                    }
                  },
                  "company": {
                    "type": "array",
                    "description": "Companies providing this service",
                    "items": {
                      "type": "object",
                      "required": [
                        "id"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "description": "Company ID"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/service/delete": {
      "delete": {
        "summary": "Delete a service",
        "tags": [
          "Service"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The service ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Service deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}